pyproject.toml
[tool.poetry]
name = "mongo-with-fastapi"
version = "0.0.1"
description = ""
authors = ["AndrewChan <1085718846@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.3"}
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
motor = "^3.6.0"
uvloop = "^0.21.0"
gunicorn = "^23.0.0"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
redis = "^5.2.1"
prometheus-client = "^0.21.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
